Factor personalisation page to enable building a summary page
