+++ src/pages/PersonalisationWizardPage.js
@@ -8,49 +8,12 @@ import mui from "material-ui";
 import reactMixin from "react-mixin";
 import _ from "underscore";
 
+import BasePersonalisationPage from './BasePersonalisationPage';
 import categories from '../constants/categories';
 import components from '../components';
 import icons from '../icons';
 
-/*::`*/@reactMixin.decorate(Router.Navigation)/*::`;*/
-/*::`*/@reactMixin.decorate(Router.State)/*::`;*/
-class PersonalisationPage extends React.Component {
-    constructor(props: Object) {
-        super(props);
-        this.state = {
-            subpage: 0,
-        };
-    }
-
-    // flow:disable
-    static childContextTypes = {
-        controller: React.PropTypes.instanceOf(PersonalisationPage),
-    };
-
-    getChildContext(): Object {
-        return {
-            controller: this,
-        };
-    }
-
-    // flow:disable not supported yet
-    get category(): categories.Category {
-        if (this._category) {
-            return this._category;
-        }
-
-        var category = _.findWhere(categories, {
-            key: this.props.params.page,
-        });
-
-        if (category === undefined) {
-            throw "No such category " + this.props.params.page;
-        }
-
-        this._category = category;
-        return category;
-    }
-
+class PersonalisationWizardPage extends BasePersonalisationPage {
     previousStep(): void {
         var subpage = this.state.subpage - 1;
 
@@ -88,4 +51,4 @@ class PersonalisationPage extends React.Component {
 
 }
 
-export default PersonalisationPage;
+export default PersonalisationWizardPage;
