+++ src/styles/bundle.scss
@@ -14,8 +14,8 @@
 @import '../components/ServicePane.scss';
 @import '../components/TransportTime.scss';
 @import '../icons/icons.scss';
+@import '../pages/BasePersonalisationPage.scss';
 @import '../pages/BrandedPage.scss';
 @import '../pages/HomePage.scss';
-@import '../pages/PersonalisationPage.scss';
 @import '../pages/ResultsListPage.scss';
 @import '../pages/ResultsMapPage.scss';
